server: mock engine clients, test error handling
