+++ server/host_test.go
@@ -1,8 +1,10 @@
 package server
 
 import (
-	"github.com/stretchr/testify/assert"
+	"fmt"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestLoadHost(t *testing.T) {
@@ -55,4 +57,26 @@ func TestLoadHostConfigClientOptions(t *testing.T) {
 	assert.NoError(t, err, "loadHost")
 	assert.Equalf(t, "test", host.String(), "Host.String()")
 	assert.Equalf(t, "private@localhost", host.client.String(), "Host.String()")
+}
+
+func TestLoadHostConfigError(t *testing.T) {
+	var engine = makeTestEngine(testConfig{})
+
+	var _, err = loadHost(engine, HostID("test"), HostConfig{
+		SNMP: "localhost:asdf",
+	})
+
+	assert.EqualErrorf(t, err, "parse \"udp+snmp://localhost:asdf\": invalid port \":asdf\" after host", "loadHost ParseConfig")
+}
+
+func TestLoadHostClientError(t *testing.T) {
+	var engine = makeTestEngine(testConfig{clientMock: true})
+
+	engine.mockClient("localhost", fmt.Errorf("Test error"))
+
+	var _, err = loadHost(engine, HostID("test"), HostConfig{
+		SNMP: "localhost",
+	})
+
+	assert.EqualErrorf(t, err, "NewClient test: Test error", "loadHost client")
 }
